Extract permissions related custom commands
