+++ frontend/test/__support__/e2e/commands.js
@@ -4,6 +4,9 @@ import "./commands/ui/icon";
 import "./commands/user/createUser";
 import "./commands/user/authentication";
 
+import "./commands/permissions/updatePermissions";
+import "./commands/permissions/sandboxTable";
+
 import "./commands/overwrites/log";
 
 Cypress.Commands.add("createDashboard", name => {
@@ -53,122 +56,6 @@ Cypress.Commands.add(
       },
       display,
       visualization_settings,
-    });
-  },
-);
-
-/**
- * PERMISSIONS
- *
- * As per definition for `PUT /graph` from `permissions.clj`:
- *
- * "This should return the same graph, in the same format,
- * that you got from `GET /api/permissions/graph`, with any changes made in the wherever necessary.
- * This modified graph must correspond to the `PermissionsGraph` schema."
- *
- * That's why we must chain GET and PUT requests one after the other.
- */
-
-Cypress.Commands.add(
-  "updatePermissionsGraph",
-  (groupsPermissionsObject = {}) => {
-    if (typeof groupsPermissionsObject !== "object") {
-      throw new Error("`groupsPermissionsObject` must be an object!");
-    }
-
-    cy.log("Fetch permissions graph");
-    cy.request("GET", "/api/permissions/graph").then(
-      ({ body: { groups, revision } }) => {
-        const UPDATED_GROUPS = Object.assign(groups, groupsPermissionsObject);
-
-        cy.log("Update/save permissions");
-        cy.request("PUT", "/api/permissions/graph", {
-          groups: UPDATED_GROUPS,
-          revision,
-        });
-      },
-    );
-  },
-);
-
-Cypress.Commands.add(
-  "updatePermissionsSchemas",
-  ({ schemas = {}, user_group = 4, database_id = 1 } = {}) => {
-    if (typeof schemas !== "object") {
-      throw new Error("`schemas` must be an object!");
-    }
-
-    cy.request("GET", "/api/permissions/graph").then(
-      ({ body: { groups, revision } }) => {
-        const UPDATED_GROUPS = Object.assign(groups, {
-          [user_group]: {
-            [database_id]: {
-              schemas,
-            },
-          },
-        });
-
-        cy.log("Update/save permissions");
-        cy.request("PUT", "/api/permissions/graph", {
-          groups: UPDATED_GROUPS,
-          revision,
-        });
-      },
-    );
-  },
-);
-
-Cypress.Commands.add("updateCollectionGraph", (groupsCollectionObject = {}) => {
-  if (typeof groupsCollectionObject !== "object") {
-    throw new Error("`groupsCollectionObject` must be an object!");
-  }
-
-  cy.log("Fetch permissions graph");
-  cy.request("GET", "/api/collection/graph").then(
-    ({ body: { groups, revision } }) => {
-      const UPDATED_GROUPS = Object.assign(groups, groupsCollectionObject);
-
-      cy.log("Update/save permissions");
-      cy.request("PUT", "/api/collection/graph", {
-        groups: UPDATED_GROUPS,
-        revision,
-      });
-    },
-  );
-});
-
-Cypress.Commands.add(
-  "sandboxTable",
-  ({
-    attribute_remappings = {},
-    card_id = null,
-    group_id = 4,
-    table_id = 2,
-  } = {}) => {
-    // Extract the name of the table, as well as `schema` and `db_id` that we'll need later on for `cy.updatePermissionsSchemas()`
-    cy.request("GET", "/api/table").then(({ body: tables }) => {
-      const { name, schema, db_id } = tables.find(
-        table => table.id === table_id,
-      );
-      const attr = Object.keys(attribute_remappings).join(", "); // Account for the possiblity of passing multiple user attributes
-
-      cy.log(`Sandbox "${name}" table on "${attr}"`);
-      cy.request("POST", "/api/mt/gtap", {
-        attribute_remappings,
-        card_id,
-        group_id,
-        table_id,
-      });
-
-      cy.updatePermissionsSchemas({
-        schemas: {
-          [schema]: {
-            [table_id]: { query: "segmented", read: "all" },
-          },
-        },
-        user_group: group_id,
-        database_id: db_id,
-      });
     });
   },
 );
